Update [R].Consumer.basket.analysis
Analysing big data, provided by a business partner - retailer.
Clustering the data. Thus, making it easier to operate with.
Analyse the behavior of the final customers based on their purchases, time of purchases, etc.
Composing a consumer basket which consists of products/ items which are suitable to an individual based on the behavoiral analysis.